Validation

+++ api/Controllers/CommentController.cs
@@ -22,6 +22,11 @@ namespace api.Controllers
         }
         [HttpGet]
         public async Task<IActionResult> GetAll() {
+
+            if(!ModelState.IsValid) {
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepo.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDto());
 
@@ -29,8 +34,13 @@ namespace api.Controllers
         }
         // [HttpGet]
         // [Route("{id}")]
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id) {
+
+            if(!ModelState.IsValid) {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if(comment == null) {
@@ -40,8 +50,13 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost ("{productId}")]
+        [HttpPost ("{productId:int}")]
         public async Task<IActionResult> Create([FromRoute] int productId, CreateCommentDto commentDto) {
+
+            if(!ModelState.IsValid) {
+                return BadRequest(ModelState);
+            }
+
             if(!await _productRepo.ProductExists(productId)) {
                 return BadRequest(" Product Does not exists");
             }
@@ -52,8 +67,13 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new {id = commentModel.Id}, commentModel.ToCommentDto());
         }
 
-        [HttpPut ("{id}")]
+        [HttpPut ("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto) {
+
+            if(!ModelState.IsValid) {
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
             if (comment == null) {
                 return NotFound("Comment not found");
@@ -61,9 +81,14 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete ([FromRoute] int id) {
-          var commentModel = await _commentRepo.DeleteAsync(id);
+
+            if(!ModelState.IsValid) {
+                return BadRequest(ModelState);
+            }
+
+            var commentModel = await _commentRepo.DeleteAsync(id);
         
             if (commentModel == null) {
                 return NotFound("Comment does not exists");
